video-automation-workflow/requirements.txt
pyyaml>=6.0
click>=8.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
edge-tts>=6.1.0
